Eris_backend_project/requirements.txt
certifi==2018.10.15
chardet==3.0.4
click==6.7
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-admin==1.3.2
django-excel-response2==2.0.8
django-filter==2.0.0
django-registration==3.0
django-six==1.0.4
django-swagger-ui==0.1.9
djangorestframework==3.9.0
drf-yasg==1.11.0
flex==6.13.2
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
Markdown==2.6.11
MarkupSafe==1.1.0
openapi==1.1.0
openapi-codec==1.3.2
psycopg2-binary==2.7.5
pydot==1.2.4
pyparsing==2.2.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
rfc3987==1.3.8
ruamel.yaml==0.15.77
screen==1.0.1
six==1.11.0
ssv==0.1.1
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
xlwt==1.3.0
